requirements.txt
absl-py==0.11.0
cycler==0.10.0
dataclasses==0.6
kiwisolver==1.3.1
matplotlib==3.3.4
mediapipe==0.8.2
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==8.1.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
termcolor==1.1.0